refactoring deletion profile
