+++ CommentMap.Mvc/Areas/Identity/Pages/Account/Manage/DeleteProfile.cshtml
@@ -1,0 +1,36 @@
+﻿@page
+@model DeletePersonalDataModel
+@{
+    ViewData["Title"] = "Delete your profile";
+    ViewData["ActivePage"] = ManageNavPages.DeleteProfile;
+}
+
+<h3>@ViewData["Title"]</h3>
+
+<div class="alert alert-warning" role="alert">
+    <p>
+        <strong>Deletion is irreversible.</strong>
+    </p>
+    <p>This operation will permanently remove your account, and this cannot be recovered.</p>
+</div>
+
+<div class="row">
+    <div class="col-md-6">
+        <form id="delete-user" method="post">
+            <div asp-validation-summary="ModelOnly" class="text-danger" role="alert"></div>
+            @if (Model.RequirePassword)
+            {
+                <div class="form-floating mb-3">
+                    <input asp-for="Input.Password" class="form-control" autocomplete="current-password" aria-required="true" placeholder="Please enter your password." />
+                    <label asp-for="Input.Password" class="form-label"></label>
+                    <span asp-validation-for="Input.Password" class="text-danger"></span>
+                </div>
+            }
+            <button class="w-100 btn btn-lg btn-danger" type="submit">Delete profile</button>
+        </form>
+    </div>
+</div>
+
+@section Scripts {
+    <partial name="_ValidationScriptsPartial" />
+}
