--- CommentMap.Mvc/Areas/Identity/Pages/Account/Manage/DownloadPersonalData.cshtml
@@ -1,12 +1,0 @@
-﻿@page
-@model DownloadPersonalDataModel
-@{
-    ViewData["Title"] = "Download Your Data";
-    ViewData["ActivePage"] = ManageNavPages.PersonalData;
-}
-
-<h3>@ViewData["Title"]</h3>
-
-@section Scripts {
-    <partial name="_ValidationScriptsPartial" />
-}
