+++ CommentMap.Mvc/Areas/Identity/Pages/Account/Manage/Index.cshtml
@@ -1,7 +1,7 @@
 ﻿@page
 @model IndexModel
 @{
-    ViewData["Title"] = "Profile";
+    ViewData["Title"] = "Edit your profile";
     ViewData["ActivePage"] = ManageNavPages.Index;
 }
